program: upgrade to anchor 0.29.0 and solana 1.16 (#829)
* sdk: account for settle lp in accoutnt breakdown

* init

* update borsh

* revert changes to adminClient

* try simplifying lifetimes

* remove unnecessary lifetimes

* remove unnecessary lifetimes

* fix idl

* fix cargo fmt

* try updating anchor/solana/rust version

* fix cargo test

* tweak switchboard test

* try using rust 1.68

* Revert "try using rust 1.68"

This reverts commit 93c952730f7f193ae977b77597a442f7a4aa67ef.

* fix pre-commit

* CHANGELOG

+++ programs/drift/src/controller/orders/tests.rs
@@ -4807,9 +4807,9 @@ pub mod fill_order {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: oracle_price.twap as i64,
-                    last_oracle_price_twap_5min: oracle_price.twap as i64,
-                    last_oracle_price: oracle_price.agg.price as i64,
+                    last_oracle_price_twap: oracle_price.twap,
+                    last_oracle_price_twap_5min: oracle_price.twap,
+                    last_oracle_price: oracle_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -5012,9 +5012,9 @@ pub mod fill_order {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: oracle_price.twap as i64,
-                    last_oracle_price_twap_5min: oracle_price.twap as i64,
-                    last_oracle_price: oracle_price.agg.price as i64,
+                    last_oracle_price_twap: oracle_price.twap,
+                    last_oracle_price_twap_5min: oracle_price.twap,
+                    last_oracle_price: oracle_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -5354,9 +5354,9 @@ pub mod fill_order {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: oracle_price.twap as i64,
-                    last_oracle_price_twap_5min: oracle_price.twap as i64,
-                    last_oracle_price: oracle_price.agg.price as i64,
+                    last_oracle_price_twap: oracle_price.twap,
+                    last_oracle_price_twap_5min: oracle_price.twap,
+                    last_oracle_price: oracle_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -9232,9 +9232,9 @@ pub mod force_cancel_orders {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: oracle_price.twap as i64,
-                    last_oracle_price_twap_5min: oracle_price.twap as i64,
-                    last_oracle_price: oracle_price.agg.price as i64,
+                    last_oracle_price_twap: oracle_price.twap,
+                    last_oracle_price_twap_5min: oracle_price.twap,
+                    last_oracle_price: oracle_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -9528,9 +9528,9 @@ pub mod get_maker_orders_info {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: pyth_price.twap as i64,
-                    last_oracle_price_twap_5min: pyth_price.twap as i64,
-                    last_oracle_price: pyth_price.agg.price as i64,
+                    last_oracle_price_twap: pyth_price.twap,
+                    last_oracle_price_twap_5min: pyth_price.twap,
+                    last_oracle_price: pyth_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -9722,9 +9722,9 @@ pub mod get_maker_orders_info {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: pyth_price.twap as i64,
-                    last_oracle_price_twap_5min: pyth_price.twap as i64,
-                    last_oracle_price: pyth_price.agg.price as i64,
+                    last_oracle_price_twap: pyth_price.twap,
+                    last_oracle_price_twap_5min: pyth_price.twap,
+                    last_oracle_price: pyth_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -9917,9 +9917,9 @@ pub mod get_maker_orders_info {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: pyth_price.twap as i64,
-                    last_oracle_price_twap_5min: pyth_price.twap as i64,
-                    last_oracle_price: pyth_price.agg.price as i64,
+                    last_oracle_price_twap: pyth_price.twap,
+                    last_oracle_price_twap_5min: pyth_price.twap,
+                    last_oracle_price: pyth_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -10098,9 +10098,9 @@ pub mod get_maker_orders_info {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: pyth_price.twap as i64,
-                    last_oracle_price_twap_5min: pyth_price.twap as i64,
-                    last_oracle_price: pyth_price.agg.price as i64,
+                    last_oracle_price_twap: pyth_price.twap,
+                    last_oracle_price_twap_5min: pyth_price.twap,
+                    last_oracle_price: pyth_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -10353,9 +10353,9 @@ pub mod get_maker_orders_info {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: pyth_price.twap as i64,
-                    last_oracle_price_twap_5min: pyth_price.twap as i64,
-                    last_oracle_price: pyth_price.agg.price as i64,
+                    last_oracle_price_twap: pyth_price.twap,
+                    last_oracle_price_twap_5min: pyth_price.twap,
+                    last_oracle_price: pyth_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -10550,9 +10550,9 @@ pub mod get_maker_orders_info {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: pyth_price.twap as i64,
-                    last_oracle_price_twap_5min: pyth_price.twap as i64,
-                    last_oracle_price: pyth_price.agg.price as i64,
+                    last_oracle_price_twap: pyth_price.twap,
+                    last_oracle_price_twap_5min: pyth_price.twap,
+                    last_oracle_price: pyth_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
